FIX: refactor project structure
